Refactor bandwidth utilities into shared crate
