+++ crates/engine/Cargo.toml
@@ -14,6 +14,7 @@ rsync-filters = { path = "../filters" }
 rsync-checksums = { path = "../checksums" }
 rsync-compress = { path = "../compress" }
 rsync-protocol = { path = "../protocol" }
+rsync-bandwidth = { path = "../bandwidth" }
 globset = "0.4"
 
 [features]
@@ -25,3 +26,4 @@ filetime = "0.2"
 tempfile = "3.10"
 rustix = { version = "0.38", features = ["fs", "process"] }
 xattr = "1.3"
+rsync-bandwidth = { path = "../bandwidth", features = ["test-support"] }
